# lib/api-zod/src/generated/types/entityState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VARUNA security analysis console API
 * OpenAPI spec version: 0.1.0
 */

export type EntityState = typeof EntityState[keyof typeof EntityState];


export const EntityState = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  skipped: 'skipped',
} as const;