# lib/api-zod/src/generated/types/patch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VARUNA security analysis console API
 * OpenAPI spec version: 0.1.0
 */
import type { PatchSource } from './patchSource';
import type { PatchStatus } from './patchStatus';

export interface Patch {
  id: string;
  findingId: string;
  title: string;
  source: PatchSource;
  status: PatchStatus;
  createdAt: string;
  originalCode: string;
  proposedCode: string;
  explanation: string;
}