# lib/api-zod/src/generated/types/protocolTest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VARUNA security analysis console API
 * OpenAPI spec version: 0.1.0
 */
import type { EntityState } from './entityState';

export interface ProtocolTest {
  id: string;
  projectName: string;
  target: string;
  fields: string[];
  strategy: string[];
  inputs: number;
  coverage: number;
  crashes: number;
  uniqueCrashes: number;
  /** @nullable */
  triggeringField?: string | null;
  state: EntityState;
  createdAt: string;
}