# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VARUNA security analysis console API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportInput {
  projectId: string;
}